fix: [qa-sweep] Flaky under load: `update::tests::stage::rollback_repla… [ORB-12174]
[qa-sweep] Flaky under load: `update::tests::stage::rollback_replaces_a_running_executable_atomically` spawns a freshly copied executable without the ETXTBSY retry ORB-11340 added elsewhere

Orchestrated-By: opus
