+++ crates/orbit-common/Cargo.toml
@@ -16,9 +16,9 @@ clap = ["orbit-types/clap"]
 # Shared SQLite connection defaults (utility::sqlite). Optional so that
 # non-SQLite consumers (orbit-policy, orbit-exec, ...) do not pull rusqlite.
 sqlite = ["dep:rusqlite"]
-# Exposes `test_fixtures` (frozen model-name constants) and `test_env` (scoped
-# environment guards) to integration tests in sibling crates, which cannot see
-# this crate's `#[cfg(test)]` items.
+# Exposes `test_fixtures` (frozen model-name constants), `test_env` (scoped
+# environment guards), and `test_process` (spawn-retry helpers) to integration
+# tests in sibling crates, which cannot see this crate's `#[cfg(test)]` items.
 test-util = []
 
 [dependencies]
