+++ crates/orbit-common/src/lib.rs
@@ -30,6 +30,9 @@ pub mod test_env;
 #[cfg(any(test, feature = "test-util"))]
 pub mod test_fixtures;
 
+#[cfg(any(test, feature = "test-util"))]
+pub mod test_process;
+
 pub use error::{
     ArtifactOrigin, ArtifactOriginMode, DependencyNotDelivered, FrictionNotLocal, NotFoundKind,
     OrbitError, RecoverableVcsConflict, WorkspaceClaimHeld,
